requirements.txt
PyQt5==5.15.10
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest==7.4.3